requirements.txt
Django==1.8.2
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.1.1
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
Pillow==2.5.3
DjangoRestless==0.0.6
django-contrib-comments
django-localflavor==1.0
newrelic
django-s3-folder-storage==0.2
stripe
django-merchant
django-cors-headers
django-s3direct
django-markdown==0.8.4
google-api-python-client
contact_importer
django-social-auth==0.7.28
python-social-auth==0.2.2
django-sslify>=0.2
django-wysiwyg
django-tinymce==1.5.2
django-ckeditor
django-pagination
python-instagram
django-gravatar2
pylibmc==1.5.0
django-pylibmc==0.6.0
opentok>=2.2
python-dateutil
pyqrcode==1.2.1
boto==2.45
pypng==0.0.18
Flask==0.10.1
twilio==5.4.0
fake-factory==0.5.3
DjangoRestless==0.0.6
djangorestframework==3.4.7
django-rest-swagger==2.1.0
django-rest-auth==0.8.2
django-allauth==0.29.0
pusher==1.6
boto3==1.4.4